docker/docker-main/requirements.txt
Flask>=2.0.1
Flask-SQLAlchemy>=2.5.1
Flask-WTF>=1.0.0
smbprotocol>=1.8.3
SQLAlchemy>=1.4.23
waitress>=1.4.4
Werkzeug>=2.0.1
WTForms>=3.0.0
psycopg2-binary>=2.8.6
flask-cors>=1.0.0
python-docx>=0.8.11
pandas>=1.1.5
xlrd>=2.0.1
samba
xlutils
